Add files via upload
This jupyter notebook defines a two-way (type-II) linear least squares regression that is ideal for geochemical stoichiometric analysis because it allows for error in the x parameter as well as the y parameter, and as such does not assume dependence between the x and y axes. The algorithm for this two-way regression was written into MATLAB (lsqfitma.m) by Ed Peltzer circa 1995, and the original MATLAB code is included as a comment in this file. 

To use the two-way regression in jupyter notebook, simply include the third cell of this notebook into your code to define a lsqfitma(x,y) function